+++ src/verify-request/utilities.ts
@@ -3,6 +3,8 @@ import {Context} from 'koa';
 import Shopify from '@shopify/shopify-api';
 
 import {Routes} from './types';
+import {AccessMode} from '../types';
+import {DEFAULT_ACCESS_MODE} from '../auth';
 
 export function redirectToAuth(
   {fallbackRoute, authRoute}: Routes,
@@ -18,9 +20,9 @@ export function redirectToAuth(
   ctx.redirect(routeForRedirect);
 }
 
-export async function clearSession(ctx: Context) {
+export async function clearSession(ctx: Context, accessMode: AccessMode = DEFAULT_ACCESS_MODE) {
   try {
-    await Shopify.Utils.deleteCurrentSession(ctx.req, ctx.res);
+    await Shopify.Utils.deleteCurrentSession(ctx.req, ctx.res, accessMode === 'online');
   }
   catch (error) {
     if (error instanceof Shopify.Errors.SessionNotFound) {